pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-gaant"
version = "0.1.0"
description = "Manage GitHub Issues as Gantt chart tasks - Sync with YAML/Excel and generate Mermaid diagrams"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Lautaro Matas", email = "lmatas@gmail.com" }
]
keywords = [
    "github",
    "gantt",
    "project-management",
    "issues",
    "mermaid",
    "yaml",
    "excel",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "PyGithub>=2.1.1",
    "gql[requests]>=3.5.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/lmatas/github-gaant"
Documentation = "https://github.com/lmatas/github-gaant#readme"
Repository = "https://github.com/lmatas/github-gaant.git"
Issues = "https://github.com/lmatas/github-gaant/issues"

[project.scripts]
gaant = "github_gaant.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["github_gaant"]

[tool.hatch.build.targets.sdist]
include = [
    "/github_gaant",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
